This is small implemetation of my ongoing project without much preprocessing.
document is selected and then applied binarization using nick algorithm technique, then the character is extracted using dfs now each character is resized and again binarised so that it will fit for neural network. Then the image is supplied to trained neural network and predicted character is printed in corresponding window.